update debug marker
